feat: 글 삭제 구현

+++ src/main/java/com/example/demo/post/controller/PostController.java
@@ -162,16 +162,20 @@ public class PostController {
      * 글 삭제
      */
     @PreAuthorize("isAuthenticated()")
-    @PostMapping("/{id}/delete")
+    @GetMapping("/{id}/delete")
     public String postDelete(
             Principal principal,
             @PathVariable("id") Long id) {
 
         Post post = postService.getPost(id);
 
-//        if(!post.getAuthor().getUsername().equals(principal.getName())) {
-//            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
+//        if (post == null) {
+//            throw new DataNotFoundException("%d번 질문은 존재하지 않습니다.");
 //        }
+
+        if (!post.getAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다.");
+        }
 
         postService.delete(post);
